JariBakken: Fix Chrome/Firefox issues on Vista (hopefully)
r7987

+++ common/test/rb/lib/selenium/webdriver/spec_support/rack_server.rb
@@ -8,13 +8,22 @@ module Selenium
         HOST = 'localhost'
         PORT = 8182
 
-        def initialize
-          @app = Rack::File.new("#{WebDriver.root}/common/src/web")
+        def initialize(path)
+          @path = path
+          @app  = Rack::File.new(path)
         end
 
         def start
-          @pid = fork { Rack::Handler::WEBrick.run(@app, :Host => HOST, :Port => PORT) }
+          if [:windows, :java].include? Platform.platform
+            @thread = Thread.new { run }
+          else
+            @pid = fork { run }
+          end
           sleep 2
+        end
+
+        def run
+          Rack::Handler::WEBrick.run(@app, :Host => HOST, :Port => PORT)
         end
 
         def where_is(file)
@@ -22,7 +31,9 @@ module Selenium
         end
 
         def stop
-          if @pid
+          if defined?(@thread) && @thread
+            @thread.kill
+          elsif defined?(@pid) && @pid
             Process.kill('KILL', @pid)
             Process.waitpid(@pid)
           end
